move ssh to its own module
